requirements.txt
flask
Flask-Script
Flask-WTF
python-jenkins
PyYaml
six
